requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-firebase-orm==0.6.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==2.13.0
gcloud==0.17.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.7
jws==0.1.3
msgpack==1.0.8
oauth2client==3.0.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.4.3
PyJWT==2.9.0
pyparsing==3.1.2
Pyrebase==3.0.27
python-dotenv==1.0.1
python-jwt==2.0.1
redis==5.0.6
requests==2.32.3
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
spotipy==2.24.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
